Extract number_input
